Add the 5x5 board and put the no-adjacent-buildings rule back on the table
Two corrections and one new lever.

Correction 1: --control-guerrero was wrong to recommend. The base rule is what
was actually asked for -- any unit takes an unoccupied building by entering it,
only the warrior enters an occupied one -- and dropping the flag improves the
balance rather than hurting it.

Correction 2: the no-adjacent-buildings rule was dropped when the king game was
built, on the grounds that buildings no longer block movement. Putting it back
on a bigger board is what the design actually wanted, so the board is now
configurable: --lado 5. Each plane became a UInt128 because 25 squares of 4 bits
is 100 bits, which does not fit in a ulong, and the move encoding widened to 5
bits per square.

5x5 with --no-pegado is the best game measured on three of four axes. The build
order finally becomes tactical -- the most-played order drops from 69.6% to
22.9% of six permutations, and the first building splits 29/38/33 instead of
14/17/69 -- draws fall to 24.2%, and the castle becomes the way to win at 67.2%
of finishes instead of 29.6%. It fails on balance: deviation 10.5, and rising
with depth.

--guerrero-veloz (the warrior charges up to two squares in a straight line)
confirms the diagnosis exactly. It leaves the five-move race untouched and only
makes interference arrive in time, and the deviation collapses from 10.5 to 1.6
and stops growing with depth. But the cure costs too much: the warrior takes 45%
of all moves, the priest falls to 5.7%, and castle finishes drop from 67% to
11%. It fixes the wrong number.

+++ CastilloDorado/rey/Partida.cs
@@ -55,7 +55,7 @@ public sealed class PoliticaBusqueda : Politica {
 /// <summary>Corre partidas enteras y lleva la cuenta de las repeticiones.</summary>
 public sealed class Mesa {
     private readonly Juego _g;
-    private readonly Dictionary<(ulong, ulong, int), int> _vistas = new();
+    private readonly Dictionary<(UInt128, UInt128, int), int> _vistas = new();
     private readonly List<Pos> _historia = new();
 
     public Mesa(Juego g) { _g = g; }
